Major change in infrastructures
I decided that this Node module is really not a Node module and rather
Node is being built to be used as OF's scripting language. Therefore
ofRunApp will remain blocking the Node's main loop. Honestly, I see no
point in letting Node's loop to remain alive. This is a graphics
application framework for god's sake! One can use Fibers or TAGG or
Clusters if they badly want the node loop to remain alive!!